adding if statement to make sure doc doesnt get deleted

+++ cla_backend/apps/reports/views.py
@@ -203,10 +203,11 @@ def download_file(request, file_name="", *args, **kwargs):
     response["Content-Disposition"] = "attachment; filename=%s" % smart_str(file_name)
     response["X-Sendfile"] = smart_str("%s%s" % (settings.TEMP_DIR, file_name))
 
-    try:
-        export_record = Export.objects.get(user_id=request.user.pk, path__endswith=file_name)
-        export_record.delete()
-    except Export.DoesNotExist:
-        raise Http404("Export does not exist")
+    if not file_name.__contains__("WEEKLY-REPORT"):
+        try:
+            export_record = Export.objects.get(user_id=request.user.pk, path__endswith=file_name)
+            export_record.delete()
+        except Export.DoesNotExist:
+            raise Http404("Export does not exist")
 
     return response
